mari_ai_agent/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Environment variables
python-dotenv==1.0.0

# AI/ML
openai==1.3.7
numpy==1.24.3
pandas==2.0.3

# HTTP client for testing
httpx==0.25.2

# Caching
redis==5.0.1

# Utilities
pydantic[email]==2.5.0